Code structure rearrangement
