+++ client/chatting.ts
@@ -4,139 +4,7 @@ import { fromHex, randomBytes, toHex } from './crypto/cryptography.js';
 import { ModeOfOperation, utils } from 'aes-js';
 import $ from 'jquery';
 import { io, type Socket } from 'socket.io-client';
-import { getCookie } from './lib.js';
-
-// Storage of all DH keypairs in local storage, each message has its own pair to ensure forward secrecy
-interface Keypair {
-	publicKey: string;
-	privateKey: string;
-};
-// Object that contains a dictionary of message IDs and keypairs that correspond to them.
-interface CodeBook {
-	messageKeypair: Map<string, Keypair>;
-};
-// Message queue object, tracks messages awaiting key from other user, messages pending encryption and messages that have been decrypted
-interface MessageQueue {
-	awaitingKey: string[];
-	pendingMessage: Map<string, string>; // MessageID, content
-	decipheredMessages: Map<string, string>; // MessageID, content
-}
-
-// Gets the user id from a hidden field in DOM
-function getUser (): string {
-	return $('#user_id')[0].innerText;
-}
-
-// Function to convert Map into JSON
-function replacer (key: any, value: any): any {
-	if (value instanceof Map) {
-		return {
-			dataType: 'Map',
-			value: Array.from(value.entries()) // or with spread: value: [...value]
-		};
-	} else {
-		return value;
-	}
-}
-
-// Function to convert JSON into map
-function reviver (key: any, value: any): any {
-	if (typeof value === 'object' && value !== null) {
-		if (value.dataType === 'Map') {
-			return new Map(value.value);
-		}
-	}
-	return value;
-}
-
-// Obtains the codebook object from local storage
-function getCodebook (): CodeBook {
-	const userId = getUser();
-	const codebookString = localStorage.getItem(userId + '-keys');
-	if (codebookString === null) {
-		// If no codebook has been saved, make a new one
-		const codebook: CodeBook = {
-			messageKeypair: new Map<string, Keypair>()
-		};
-		return codebook;
-	} else {
-		const codebook: CodeBook = JSON.parse(codebookString, reviver);
-		return codebook;
-	}
-}
-
-// Obtains the message queue from local storage
-function getMessageQueue (): MessageQueue {
-	const userId = getUser();
-	const queueString = localStorage.getItem(userId + '-queue');
-	if (queueString === null) {
-		// If no queue exists, make a new one
-		const queue: MessageQueue = {
-			awaitingKey: [],
-			pendingMessage: new Map<string, string>(),
-			decipheredMessages: new Map<string, string>()
-		};
-		return queue;
-	} else {
-		const queue: MessageQueue = JSON.parse(queueString, reviver) as MessageQueue;
-		return queue;
-	}
-}
-
-// Enqueues a message in the message queue
-function queueMessage (messageId: string, messageContent: string): void {
-	const queue = getMessageQueue();
-	queue.pendingMessage.set(messageId, messageContent);
-	localStorage.setItem(getUser() + '-queue', JSON.stringify(queue, replacer));
-}
-
-// Dequeues a message in the message queue
-function dequeueMessage (messageId: string): void {
-	const queue = getMessageQueue();
-	queue.pendingMessage.delete(messageId);
-	localStorage.setItem(getUser() + '-queue', JSON.stringify(queue, replacer));
-
-	const messageList = document.getElementById('messages');
-	if (messageList !== null) location.reload(); // reload the chat page so the new messages are visible
-}
-
-// Marks down a message as awaiting key
-function awaitKey (messageId: string): void {
-	const queue = getMessageQueue();
-	queue.awaitingKey.push(messageId);
-	localStorage.setItem(getUser() + '-queue', JSON.stringify(queue, replacer));
-}
-
-// Resolves a message awaiting key
-function stopAwaitKey (messageId: string): void {
-	const queue = getMessageQueue();
-	const index = queue.awaitingKey.indexOf(messageId);
-	if (index > -1) {
-		// if a message is found in the awaiting key array, remove it
-		queue.awaitingKey.splice(index, 1);
-	}
-	localStorage.setItem(getUser() + '-queue', JSON.stringify(queue, replacer));
-}
-
-// Saves a deciphered message
-function saveMessage (messageId: string, messageContent: string): void {
-	const queue = getMessageQueue();
-	queue.decipheredMessages.set(messageId, messageContent);
-	localStorage.setItem(getUser() + '-queue', JSON.stringify(queue, replacer));
-}
-
-// Saves a keypair in the codebook
-function saveKeypair (messageId: string, publicKey: string, privateKey: string): void {
-	const codebook = getCodebook();
-	codebook.messageKeypair.set(messageId, { publicKey, privateKey });
-	localStorage.setItem(getUser() + '-keys', JSON.stringify(codebook, replacer));
-}
-
-// Retrieves a keypair from the codebook
-function loadKeypair (messageId: string): Keypair | undefined {
-	const codebook = getCodebook();
-	return codebook.messageKeypair.get(messageId);
-}
+import { awaitKey, dequeueMessage, getCookie, getMessageQueue, getUser, loadKeypair, queueMessage, saveKeypair, saveMessage, stopAwaitKey } from './lib.js';
 
 // Sends the ANNOUNCE chat message component to server
 function sendAnnounce (recipientId: string, messageContent: string): void {
